==> notebooks/step-1_find_words_wikitionary.ipynb <==
## 1.0 Imports

# files handling
import gzip
import json

# regular expressions
import re

# system handling
from pathlib import Path
from typing import Dict, List
from IPython.display import display
from IPython.display import Markdown as md

# string and string emcoding
from unicodedata import category, normalize

# data handling
import pandas as pd

# Arabic tools
from camel_tools.morphology.analyzer import Analyzer
from camel_tools.morphology.database import MorphologyDB
from pyarabic.araby import (
    DIACRITICS,
    LETTERS,
    SHADDA,
    is_arabicrange,
    is_vocalizedtext,
    name,
)
from pyarabic.trans import convert as translit

## 2.0 Define Paths

# Paths for reindexed Wikitionary and the Arabic words.

# Arabic words csv
data_dir = Path("../data/interim/gf_wordnet")
ar_words_gf_path = sorted(list(data_dir.glob("*.csv")))

# Path for preproceed Wikitionary
wiktionary_dir = Path("../data/processed/wikidata")
wiki_jsonl_path = wiktionary_dir / "ar-wiktextract-data.json.gz"
wiki_indices_path = wiktionary_dir / "ar_reindex.json.gz"

# Path to save words status
output_dir = Path("../data/interim/ambiguous")
output_unambiguous_dir = Path("../data/interim/unambiguous")

gf_lin_fun_path = "../data/processed/gf/WordNetAra.gf"

ar_words_gf_path

## 3.0 Define Variables

DIACRITICS = set(DIACRITICS)  # Aabic diacritics/short vowels

entites = "_".join(sorted({p for path in ar_words_gf_path for p in path.stem.split("_")[1:-3]}, reverse=True))
TIMESTAMP = ar_words_gf_path[-1].stem.split("_")[0]

entites, TIMESTAMP

SET_BASE_POS = {"N", "V", "V2", "V3", "A"}
SET_PROPN_POS = {"NP", "LN", "GN", "SN", "PN"}

# Remove last Harakat from an Arabic Word
NOT_LETTERS_PATTERN = f"[^{LETTERS}]"
DIACRITICS_PATTERN = "".join(DIACRITICS)
LAST_HARAKAT_PATTERN = re.compile(rf"[{DIACRITICS_PATTERN}](?={NOT_LETTERS_PATTERN}*$)", re.UNICODE)

pd.set_option("display.max_colwidth", None)

## 4.0 Define Functions

### 4.1 Utilities Functions

def is_nonspace_mark(x: str) -> bool:
    """detect diacritics in a string"""
    return bool(category(x) == "Mn")

def reorder_shadda(ar_string: str) -> str:
    """unicodedata.normalize put shadda before diacritics; not correct"""
    list_ar_str = list(ar_string)

    for i in range(len(list_ar_str) - 1):
        char = list_ar_str[i]
        next_char = list_ar_str[i + 1]

        if char in DIACRITICS and next_char == SHADDA:
            list_ar_str[i], list_ar_str[i + 1] = (
                next_char,
                char,
            )  # Swap shadda and diacritic

    return "".join(list_ar_str)

def remove_diacritics_ar(text: str) -> str:
    """remove diacritics in a string"""
    return "".join([t for t in text if not is_nonspace_mark(t)])

def normalize_ar(ar_vocalized: str, verbose: bool = False) -> str:
    """get the normal form for the Unicode string unistr using NFC then fix the shadda order issue"""
    if not isinstance(ar_vocalized, str):
        return ar_vocalized
    ar_norm = "".join(c for c in ar_vocalized if is_arabicrange(c)).strip()
    ar_norm = normalize("NFC", ar_norm)
    ar_norm = reorder_shadda(ar_norm)
    if verbose:
        print([name(char) for char in ar_norm])
    return ar_norm

def join_ar_words(words):
    if not isinstance(words, str):
        return words
    # join arabic lettera names by |
    # if the next word is haraka name join by -
    joined = name(words[0])
    for word in words[1:]:
        # Check if the word contains Arabic diacritics
        if word in DIACRITICS:
            joined += "-" + name(word)
        else:
            joined += " | " + name(word)
    return joined

### 4.2 Main Functions

def get_word_dictionaries(ar_word: str, wiki_idxs: Dict[str, int], wikitionary: List[dict]) -> List[dict]:
    """Get the Wikitionary info for a word"""
    word_idx = wiki_idxs.get(ar_word, [])
    word_data = [wikitionary[idx] for idx in word_idx]
    return word_data

def count_entry(ar_word, wiki_idxs, wikitionary):
    """Count howmay entry are there in the Wikitionary for a word"""
    ar_word_unvocalized = remove_diacritics_ar(ar_word)
    word_wikidicts = get_word_dictionaries(ar_word_unvocalized, wiki_idxs, wikitionary)
    return len(word_wikidicts)

def get_canonical_form(word, wiki_idxs, wikitionary):
    """get all vocalize forms of the word from Wikitionary"""
    # get all word'd entries in Wikitionary
    # Vocalized form is found in `'forms': [{'form': WORD_VOCALIZED, 'tags': ['canonical']}`
    form_vocal = []
    pos = []
    glosses = []
    word_idxs = []
    for word_idx in wiki_idxs.get(word, []):
        for _, word_data in wikitionary[word_idx].items():
            for form_dict in word_data.get("forms", []):
                if form_dict.get("tags", [None])[0] == "canonical":
                    form_word = form_dict.get("form")
                    if form_word and is_vocalizedtext(form_word):
                        form_vocal.append(normalize_ar(form_word))
                        pos.append(word_data["pos"])
                        word_idxs.append(word_idx)
                        glosses.append(
                            [
                                " / ".join(
                                    [", ".join(sense["glosses"]) for sense in word_data.get("senses", []) if "glosses" in sense]
                                )
                            ]
                        )

    if not form_vocal:
        # Not found in WikiData Use Camel Tools
        db = MorphologyDB.builtin_db()
        analyzer = Analyzer(db)
        analyses = analyzer.analyze(word)
        for analysis in analyses:
            if not form_vocal or (analysis["diac"] != form_vocal[-1] and analysis["pos"] != pos[-1]):
                form_vocal.append(LAST_HARAKAT_PATTERN.sub("", analysis["diac"]))
                pos.append(analysis["pos"])
                glosses.append(" / ".join(analysis["gloss"].split(";")))
                word_idxs.append(-1)

    return (form_vocal, pos, word_idxs, glosses) if form_vocal else (word, pd.NA, -1, pd.NA)

## 5.0 Check Wikitionary

- Get the missing words in Wikitionary
- Confirm that each Arabic word has only one entry in the Wikitionary file. If not, the short vowels need to be added or edited in the source
file to remove any disambguation.
- Export words with thier number of entries to be modified manually later

### 5.1 Load Files

with open(gf_lin_fun_path, encoding="utf-8", mode="rt") as obj_file:
    list_lin_fn_words = obj_file.read().split("\n")[3:-1]

dict_lin_words = {"en_entry": [], "ar": []}
for str_lin_fn_words in list_lin_fn_words:
    en_entry, ar_entry = str_lin_fn_words.split(" = ")
    en_entry = en_entry[4:]
    ar_word = remove_diacritics_ar(ar_entry[1:-1].split("_")[0])
    dict_lin_words["en_entry"].append(en_entry)
    dict_lin_words["ar"].append(ar_word)

df_ar_gf_old = pd.DataFrame(dict_lin_words)

df_ar_gf_all = []
for path in ar_words_gf_path:
    df = pd.read_csv(path, delimiter="\t", index_col=0).reset_index()
    df.drop("index", axis=1, inplace=True)
    df_ar_gf_all.append(df)

df_ar_gf_all = pd.concat(df_ar_gf_all, ignore_index=True).drop("en", axis=1)

df_ar_gf_old.head(5)

df_ar_gf_all.head(5)

df_merged = df_ar_gf_all.merge(df_ar_gf_old, on="en_entry", how="outer", suffixes=["", "_"], indicator=True)

df_merged.head(5)

merge_status = df_merged["_merge"].value_counts()
merge_status

if "right_only" in merge_status:
    display(df_merged[df_merged["_merge"] == "right_only"])


df_ar_gf_all = df_merged.copy()[df_merged["_merge"] != "both"]
df_ar_gf_all.drop(["_merge", "ar_"], inplace=True, axis=True)

df_ar_gf_all.shape

set_all_pos = set(df_ar_gf_all["pos"].to_list())
set_all_pos

set_other_pos = set_all_pos.difference(SET_BASE_POS).difference(SET_PROPN_POS)
set_other_pos

df_ar_gf = df_ar_gf_all.copy()[(df_ar_gf_all["pos"].isin(SET_BASE_POS)) | (df_ar_gf_all["pos"] == "MU")]
df_ar_cn_gf = df_ar_gf_all.copy()[(df_ar_gf_all["pos"].str.startswith("CN")) | (df_ar_gf_all["pos"].str.startswith("VP"))]
df_ar_pn_gf = df_ar_gf_all.copy()[df_ar_gf_all["pos"].isin(SET_PROPN_POS)]

df_ar_gf_all.shape[0], df_ar_gf.shape[0] + df_ar_cn_gf.shape[0] + df_ar_pn_gf.shape[0]

display(df_ar_gf.head(4))

# 2. load the wikitionary
with gzip.open(wiki_jsonl_path, "rt", encoding="utf-8") as gzip_obj:
    wikitionary = []
    for i, json_line in enumerate(gzip_obj):
        wikitionary.append(dict(json.loads(json_line)))
        print(f"Reading line-{i}", end="\r")

# 3. load Indices
with gzip.open(wiki_indices_path, "rt", encoding="utf-8") as gzip_obj:
    wiki_idxs = json.load(gzip_obj)

### 5.2 Get the Number of Entry for each Word

# Normalize words, the check the number of entry
df_ar_gf["entry_num"] = df_ar_gf["ar"].apply(normalize_ar).apply(count_entry, args=(wiki_idxs, wikitionary))

df_ar_gf.head()

max_entrs = df_ar_gf["entry_num"].max()
md(f"> Some words has more than {max_entrs} entries!!")

pd.DataFrame({"count": df_ar_gf["entry_num"].value_counts()})

### Add All Vocalized Forms for each Word

df_ambiguous_grouped = df_ar_gf.copy()
(
    df_ambiguous_grouped["vocal_forms"],
    df_ambiguous_grouped["wiki_pos"],
    df_ambiguous_grouped["wiki_idx"],
    df_ambiguous_grouped["senses"],
) = zip(
    *df_ambiguous_grouped["ar"]
    .apply(normalize_ar)
    .apply(remove_diacritics_ar)
    .apply(get_canonical_form, args=(wiki_idxs, wikitionary))
)

df_ambiguous_grouped.head(30)

df_ar_gf_ambiguous = df_ambiguous_grouped.copy().explode(["vocal_forms", "wiki_idx", "wiki_pos", "senses"])

df_ar_gf_ambiguous.head(5)

df_ar_gf_ambiguous.shape

## 6.0 Export Ambiguous Words

- Manually fixing the diacritics of the words that has more than one entry.
- Manually checking the missing word in Wkitionary

# Sort by count and index
df_ar_gf_ambiguous = (
    df_ar_gf_ambiguous.copy()
    .reset_index(drop=False)
    .sort_values(by=["entry_num", "en_entry", "wiki_idx", "ar"], kind="mergesort")
)

# Add Letter Names in Arabic
df_ar_gf_ambiguous["ar_letters"] = df_ar_gf_ambiguous["vocal_forms"].apply(normalize_ar).apply(join_ar_words)

df_ar_gf_ambiguous["tim_translit"] = df_ar_gf_ambiguous["vocal_forms"].apply(
    lambda x: translit(x, "arabic", "tim") if isinstance(x, str) else x
)

df_ar_gf_ambiguous["select"] = 0

# Rearrange columns order
neworder = [
    "wiki_idx",
    "no_en_entry",
    "en_entry",
    "ar",
    "vocal_forms",
    "pos",
    "wiki_pos",
    "select",
    "entry_num",
    "ar_letters",
    "senses",
    "tim_translit",
    "file"
]
df_ar_gf_ambiguous = df_ar_gf_ambiguous.reindex(columns=neworder)

display(df_ar_gf_ambiguous)

df_ar_gf_ambiguous.to_csv(output_dir / f"{TIMESTAMP}_{entites}_word_ambiguous.csv", sep="\t")
df_ar_gf_ambiguous.to_csv(output_unambiguous_dir / f"{TIMESTAMP}_{entites}_word_unambiguous.csv", sep="\t")

> Export Other DataFrames

df_ar_cn_gf.to_csv(output_dir / f"{TIMESTAMP}_{entites}_CN_ambiguous.csv", sep="\t")
df_ar_pn_gf.to_csv(output_dir / f"{TIMESTAMP}_{entites}_PN_ambiguous.csv", sep="\t")
