==> notebooks/step-2_get_morpho_features.ipynb <==
# 1.0 Imports

import gzip
import json
import re
import time
from datetime import datetime
from pathlib import Path
from typing import Dict, List
from unicodedata import normalize

import pandas as pd
from pyarabic.araby import DIACRITICS, SHADDA, is_arabicword, name, vocalizedlike

# 2.0 Define Paths

# Path for preproceed Wikitionary
wiktionary_dir = Path("../data/processed/wikidata")
wiki_jsonl_path = wiktionary_dir / "ar-wiktextract-data.json.gz"
wiki_indices_path = wiktionary_dir / "ar_reindex.json.gz"

# Arabic GF-Wordnet
data_dir = Path("../data/interim/unambiguous")
ar_words_gf_path = data_dir / "20240202.1101_Q79_Q34_Q16_word_unambiguous.csv"

# Path to save words status
output_dir = Path("../data/interim/lexicon")
output_dir.mkdir(parents=True, exist_ok=True)

# 3.0 Define Variables

DIACRITICS = set(DIACRITICS)  # Aabic diacritics/short vowels

TIMESTAMP = "_".join(ar_words_gf_path.stem.split("_")[:-2])

## 4.0 Define Functions

### 4.1 Utilities Functions

def reorder_shadda(ar_string: str) -> str:
    """unicodedata.normalize put shadda before diacritics; not correct"""
    list_ar_str = list(ar_string)

    for i in range(len(list_ar_str) - 1):
        char = list_ar_str[i]
        next_char = list_ar_str[i + 1]

        if char in DIACRITICS and next_char == SHADDA:
            list_ar_str[i], list_ar_str[i + 1] = (
                next_char,
                char,
            )  # Swap shadda and diacritic

    return "".join(list_ar_str)

def normalize_ar(ar_vocalized: str, verbose: bool = False) -> str:
    """get the normal form for the Unicode string unistr using NFC then fix the shadda order issue"""
    ar_norm = normalize("NFC", ar_vocalized)
    ar_norm = reorder_shadda(ar_norm)
    if verbose:
        print([name(char) for char in ar_norm])
    return ar_norm

## 4.2 Morphology Features

### 4.2.1 Gender

def get_noun_gender(entry, form):
    for _, word_data in entry.items():
        categories: List[str] = word_data.get("categories", [])
        for category in categories:
            masculine = category.find("masculine") != -1
            feminine = category.find("feminine") != -1
            if masculine and feminine:
                return "N"
            elif masculine:
                return "masc"
            elif feminine:
                return "fem"
        return pd.NA

def get_adj_gender(entry, form):
    for _, word_data in entry.items():
        word_forms: List[dict] = word_data.get("forms", [])
        masculine, feminine = None, None
        for form_dict in word_forms:
            if {'indefinite', 'masculine', 'singular', 'nominative'}.issubset(form_dict.get("tags", [])):
                masculine = normalize_ar(form_dict["form"])
            elif {'indefinite', 'feminine', 'singular', 'nominative'}.issubset(form_dict.get("tags", [])):
                feminine = normalize_ar(form_dict["form"])
            if masculine and feminine:
                break

        # TODO: Confusing = redesign
        if not masculine and not feminine:
            return pd.NA, pd.NA
        elif not masculine or not feminine:
            return "masc", pd.NA if masculine else "fem", pd.NA

        if vocalizedlike(form, masculine):
            return "masc", feminine
        return "fem", masculine


def get_gender(row, wikitionary):
    wiki_idx = int(row["wiki_idx"])
    pos = row["wiki_pos"]
    form: str = normalize_ar(row["vocal_forms"])
    if wiki_idx > -1:
        entry = wikitionary[wiki_idx]
        if pos == "noun":
            return get_noun_gender(entry, None)
        elif pos == "adj":
            return get_adj_gender(entry, form)
    return [pd.NA] * 2

### 4.2.2 Plural

def get_noun_plural(entry):
    for _, word_data in entry.items():
        word_forms = word_data.get("forms", [])
        for form_dict in word_forms:
            if "plural" == form_dict.get("tags", [None])[0]:
                return form_dict["form"]
        inflections = word_data.get("head_templates", [])
        for dict_inflection in inflections:
            plural = dict_inflection.get("args", {}).get("pl", None)
            if plural is not None:
                if is_arabicword(plural):
                    return normalize_ar(plural)
                elif plural == "-":
                    return normalize_ar("غير معدود")
                return normalize_ar("جمع سالم")
    return pd.NA

def get_adj_plural(entry):
    for _, word_data in entry.items():
        word_forms = word_data.get("forms", [])
        masc_pl, fem_pl = None, None
        for form_dict in word_forms:
            if ['masculine', 'plural'] == form_dict.get("tags", [None]):
                masc_pl = form_dict["form"]
            elif ['feminine', 'plural'] == form_dict.get("tags", [None]):
                fem_pl = form_dict["form"]
            if masc_pl and fem_pl:
                break
    masc_pl = normalize_ar(masc_pl) if masc_pl else pd.NA
    fem_pl = normalize_ar(fem_pl) if fem_pl else pd.NA

    return masc_pl, fem_pl

def get_plural(row, wikitionary):
    wiki_idx = int(row["wiki_idx"])
    pos = row["wiki_pos"]
    if wiki_idx > -1:
        entry = wikitionary[wiki_idx]
        if pos == "noun":
            return get_noun_plural(entry)
        elif pos == "adj":
            return get_adj_plural(entry)
    if pos == "noun":
        return pd.NA
    elif pos == "adj":
        return [pd.NA]*2

### 4.2.3 Root

def get_root(row, wikitionary):
    wiki_idx = int(row["wiki_idx"])
    if wiki_idx > -1:
        entry = wikitionary[wiki_idx]
        for _, word_data in entry.items():
            for etymology in word_data.get('etymology_templates', []):
                if etymology.get("name", "").startswith("ar-root"):
                    if len(etymology["args"]) == 3:
                        return normalize_ar("".join(list(etymology["args"].values())))
                    if len(etymology["args"]) == 1:
                        return normalize_ar("".join(etymology["args"]["1"].split()))
        return pd.NA

### 4.2.4 Verb Forms

def get_verb_form(row, wikitionary):
    wiki_idx = int(row["wiki_idx"])
    if wiki_idx > -1:
        entry = wikitionary[wiki_idx]
        form_number = None
        for _, word_data in entry.items():
            form_numbers = []
            categories: List[str] = word_data.get("categories", [])
            for category in categories:
                if match := re.search(r"(?<=form-)\w+(?= verbs)", category):
                    form_numbers.append(match.group())
            if form_numbers:
                form_number = max(form_numbers)

        return f"Form{form_number}" if form_number is not None else pd.NA


def get_imperfect_form(row, wikitionary):
    wiki_idx = int(row["wiki_idx"])
    if wiki_idx > -1:
        entry = wikitionary[wiki_idx]
        for _, word_data in entry.items():
            word_forms: List[dict] = word_data.get("forms", [])
            for form_dict in word_forms:
                if {
                    "active",
                    "indicative",
                    "masculine",
                    "non-past",
                    "imperfective",
                    "singular",
                    "third-person",
                }.issubset(form_dict.get("tags", [])):
                    return normalize_ar(form_dict["form"])

        return pd.NA


# 5.0 Load Files

# 1. load the wikitionary
with gzip.open(wiki_jsonl_path, "rt", encoding="utf-8") as gzip_obj:
    wikitionary = []
    for i, json_line in enumerate(gzip_obj):
        wikitionary.append(dict(json.loads(json_line)))
        print(f"Reading line-{i}", end="\r")

# 2. load the csv arabic-english_gf words
df_ar_gf = pd.read_csv(ar_words_gf_path, delimiter="\t", index_col=0)

df_ar_gf = df_ar_gf[df_ar_gf["select"] == 1]

# Get nouns, adjectives and verbs
df_ar_gf_nouns = df_ar_gf.copy()[(df_ar_gf["wiki_pos"] == "noun") & (df_ar_gf["select"] == 1)]
df_ar_gf_adjs = df_ar_gf.copy()[(df_ar_gf["wiki_pos"] == "adj") & (df_ar_gf["select"] == 1)]
df_ar_gf_verbs = df_ar_gf.copy()[(df_ar_gf["wiki_pos"] == "verb") & (df_ar_gf["select"] == 1)]

# 6.0 Get Morphological Features

## 6.1 Nouns

df_ar_gf_nouns["gender"] = df_ar_gf_nouns.apply(get_gender, args=(wikitionary,), axis="columns")
df_ar_gf_nouns["gender"].value_counts()

df_ar_gf_nouns["plural"] = df_ar_gf_nouns.apply(get_plural, args=(wikitionary,), axis="columns")
df_ar_gf_nouns["plural"].isna().value_counts()

df_ar_gf_nouns["root"] = df_ar_gf_nouns.apply(get_root, args=(wikitionary,), axis="columns")
df_ar_gf_nouns["root"].isna().value_counts()

## 6.2 Adjectives

df_ar_gf_adjs["gender"], df_ar_gf_adjs["other_gender_form"] = zip(
    *df_ar_gf_adjs.apply(get_gender, args=(wikitionary,), axis="columns")
)
df_ar_gf_adjs["gender"].value_counts()

df_ar_gf_adjs["masc_pl"], df_ar_gf_adjs["fem_pl"]  = zip(*df_ar_gf_adjs.apply(get_plural, args=(wikitionary,), axis="columns"))
print(df_ar_gf_adjs["masc_pl"].isna().value_counts())
print(df_ar_gf_adjs["fem_pl"].isna().value_counts())

df_ar_gf_adjs["root"] = df_ar_gf_adjs.apply(get_root, args=(wikitionary,), axis="columns")
df_ar_gf_adjs["root"].isna().value_counts()

## 6.3 Verbs

df_ar_gf_verbs["verb_form"] = df_ar_gf_verbs.apply(get_verb_form, args=(wikitionary,), axis="columns")
df_ar_gf_verbs["verb_form"].isna().value_counts()

df_ar_gf_verbs["root"] = df_ar_gf_verbs.apply(get_root, args=(wikitionary,), axis="columns")
df_ar_gf_verbs["root"].isna().value_counts()

df_ar_gf_verbs["imperfect"] = df_ar_gf_verbs.apply(get_imperfect_form, args=(wikitionary,), axis="columns")
df_ar_gf_verbs["imperfect"].isna().value_counts()

# 7.0 Export Data

# df_ar_gf_verbs.drop(columns=["select", "entry_num", "pos"], inplace=True)
df_ar_gf_nouns.drop(columns=["select", "entry_num", "pos"], inplace=True)
df_ar_gf_adjs.drop(columns=["select", "entry_num", "pos"], inplace=True)

# df_ar_gf_verbs.to_csv(output_dir / f"{TIMESTAMP}_verbs_lexicon.csv", decimal=",")
df_ar_gf_nouns.to_csv(output_dir / f"{TIMESTAMP}_nouns_lexicon.csv", decimal=",")
df_ar_gf_adjs.to_csv(output_dir / f"{TIMESTAMP}_adjectives_lexicon.csv", decimal=",")

# 8.0 Visualize Data

df_ar_gf_verbs

df_ar_gf_adjs

df_ar_gf_nouns
